+++ TinkoffASDKUI/TinkoffASDKUI/Modules/TinkoffPaySheet/Assembly/Router/ITinkoffPaySheetRouter.swift
@@ -1,0 +1,12 @@
+//
+//  ITinkoffPaySheetRouter.swift
+//  TinkoffASDKUI
+//
+//  Created by r.akhmadeev on 13.03.2023.
+//
+
+import Foundation
+
+protocol ITinkoffPaySheetRouter {
+    func openTinkoffPayLanding(completion: VoidBlock?)
+}