+++ TinkoffASDKUI/TinkoffASDKUI/SDK/AcquiringUISDK.swift
@@ -191,6 +191,7 @@ public class AcquiringUISDK: NSObject {
     private let sbpQrAssembly: ISBPQrAssembly
     private let cardListAssembly: ICardListAssembly
     private let recurrentPaymentAssembly: IRecurrentPaymentAssembly
+    private let tinkoffPaySheetAssembly: ITinkoffPaySheetAssembly
     private let tinkoffPayLandingAssembly: ITinkoffPayLandingAssembly
 
     // MARK: Init
@@ -299,8 +300,13 @@ public class AcquiringUISDK: NSObject {
         )
 
         tinkoffPayAssembly = TinkoffPayAssembly(coreSDK: coreSDK, configuration: uiSDKConfiguration)
-
         tinkoffPayLandingAssembly = TinkoffPayLandingAssembly(authChallengeService: webViewAuthChallengeService)
+
+        tinkoffPaySheetAssembly = TinkoffPaySheetAssembly(
+            coreSDK: coreSDK,
+            tinkoffPayAssembly: tinkoffPayAssembly,
+            tinkoffPayLandingAssembly: tinkoffPayLandingAssembly
+        )
 
         mainFormAssembly = MainFormAssembly(
             coreSDK: coreSDK,
@@ -1801,6 +1807,16 @@ public extension AcquiringUISDK {
             failureDelegate: failureDelegate,
             moduleCompletion: completion
         )
+
+        presentingViewController.present(viewController, animated: true)
+    }
+
+    func presentTinkoffPay(
+        on presentingViewController: UIViewController,
+        paymentFlow: PaymentFlow,
+        completion: PaymentResultCompletion? = nil
+    ) {
+        let viewController = tinkoffPaySheetAssembly.tinkoffPaySheet(paymentFlow: paymentFlow, completion: completion)
 
         presentingViewController.present(viewController, animated: true)
     }