+++ CHANGELOG.md
@@ -40,6 +40,7 @@
 * [MIC-8026] Integrate CardsController to CardList module
 * [MIC-8050] Change add card icon
 * [MIC-8095] TinkoffPay - integrate to MainForm
+* [MIC-8096] TinkoffPay - integrate with independent button
 
 ## [2.17.0] - 2023-02-13Z